requirements.txt
Flask==1.1.1
flask_session==0.3.1
pymongo==3.9.0
python_bcrypt==0.3.2
gunicorn==20.0.0
pytest==5.3.0
google-auth==1.8.1
requests==2.22.0